backend/requirements.txt
fastapi
uvicorn
qdrant-client==1.16.0
sentence-transformers
python-dotenv
numpy
google-generativeai

# Optional improvements for ingredient extraction
spacy
